requirements.txt
ultralytics==8.2.0
opencv-python==4.9.0.80
numpy==1.26.4
PyQt5==5.15.10
PyYAML==6.0.1
python-dotenv==1.0.1
requests==2.31.0
lapx>=0.5.2